Support multiple oldnames for formulae & support formula renames in API

+++ Library/Homebrew/migrator.rb
@@ -13,41 +13,34 @@ class Migrator
 
   # Error for when a migration is necessary.
   class MigrationNeededError < RuntimeError
-    def initialize(formula)
+    def initialize(oldname, newname)
       super <<~EOS
-        #{formula.oldname} was renamed to #{formula.name} and needs to be migrated by running:
-          brew migrate #{formula.oldname}
+        #{oldname} was renamed to #{newname} and needs to be migrated by running:
+          brew migrate #{oldname}
       EOS
-    end
-  end
-
-  # Error for when a formula does not replace another formula.
-  class MigratorNoOldnameError < RuntimeError
-    def initialize(formula)
-      super "#{formula.name} doesn't replace any formula."
     end
   end
 
   # Error for when the old name's path does not exist.
   class MigratorNoOldpathError < RuntimeError
-    def initialize(formula)
-      super "#{HOMEBREW_CELLAR/formula.oldname} doesn't exist."
+    def initialize(oldname)
+      super "#{HOMEBREW_CELLAR/oldname} doesn't exist."
     end
   end
 
   # Error for when a formula is migrated to a different tap without explicitly using its fully-qualified name.
   class MigratorDifferentTapsError < RuntimeError
-    def initialize(formula, tap)
+    def initialize(formula, oldname, tap)
       msg = if tap.core_tap?
-        "Please try to use #{formula.oldname} to refer to the formula.\n"
+        "Please try to use #{oldname} to refer to the formula.\n"
       elsif tap
-        "Please try to use fully-qualified #{tap}/#{formula.oldname} to refer to the formula.\n"
+        "Please try to use fully-qualified #{tap}/#{oldname} to refer to the formula.\n"
       end
 
       super <<~EOS
-        #{formula.name} from #{formula.tap} is given, but old name #{formula.oldname} was installed from #{tap || "path or url"}.
+        #{formula.name} from #{formula.tap} is given, but old name #{oldname} was installed from #{tap || "path or url"}.
         #{msg}To force migration, run:
-          brew migrate --force #{formula.oldname}
+          brew migrate --force #{oldname}
       EOS
     end
   end
@@ -65,13 +58,13 @@ class Migrator
   attr_reader :old_pin_record
 
   # Path to oldname opt.
-  attr_reader :old_opt_record
+  attr_reader :old_opt_records
 
-  # Oldname linked keg.
-  attr_reader :old_linked_keg
+  # Oldname linked kegs.
+  attr_reader :old_linked_kegs
 
-  # Path to oldname's linked keg.
-  attr_reader :old_linked_keg_record
+  # Oldname linked kegs that were fully linked.
+  attr_reader :old_full_linked_kegs
 
   # Tabs from oldname kegs.
   attr_reader :old_tabs
@@ -97,53 +90,64 @@ class Migrator
   # Path to newname keg that will be linked if old_linked_keg isn't nil.
   attr_reader :new_linked_keg_record
 
+  def self.oldnames_needing_migration(formula)
+    formula.oldnames.select do |oldname|
+      oldname_rack = HOMEBREW_CELLAR/oldname
+      next false if oldname_rack.symlink?
+      next false unless oldname_rack.directory?
+
+      true
+    end
+  end
+
   def self.needs_migration?(formula)
-    oldname = formula.oldname
-    return false unless oldname
-
-    oldname_rack = HOMEBREW_CELLAR/oldname
-    return false if oldname_rack.symlink?
-    return false unless oldname_rack.directory?
-
-    true
+    !oldnames_needing_migration(formula).empty?
   end
 
   def self.migrate_if_needed(formula, force:, dry_run: false)
-    return unless Migrator.needs_migration?(formula)
+    oldnames = Migrator.oldnames_needing_migration(formula)
+    return if oldnames.empty?
 
     begin
       if dry_run
-        ohai "Would migrate #{formula.oldname} to #{formula.name}"
+        ohai "Would migrate #{oldnames.to_sentence} to #{formula.name}"
         return
       end
-      migrator = Migrator.new(formula, force: force)
-      migrator.migrate
+
+      oldnames.each do |oldname|
+        migrator = Migrator.new(formula, oldname, force: force)
+        migrator.migrate
+      end
     rescue => e
       onoe e
     end
   end
 
-  def initialize(formula, force: false)
-    @oldname = formula.oldname
+  def initialize(formula, oldname, force: false)
+    @oldname = oldname
     @newname = formula.name
-    raise MigratorNoOldnameError, formula unless oldname
 
     @formula = formula
-    @old_cellar = HOMEBREW_CELLAR/formula.oldname
-    raise MigratorNoOldpathError, formula unless old_cellar.exist?
+    @old_cellar = HOMEBREW_CELLAR/oldname
+    raise MigratorNoOldpathError, oldname unless old_cellar.exist?
 
     @old_tabs = old_cellar.subdirs.map { |d| Tab.for_keg(Keg.new(d)) }
     @old_tap = old_tabs.first.tap
 
-    raise MigratorDifferentTapsError.new(formula, old_tap) if !force && !from_same_tap_user?
+    raise MigratorDifferentTapsError.new(formula, oldname, old_tap) if !force && !from_same_tap_user?
 
     @new_cellar = HOMEBREW_CELLAR/formula.name
     @new_cellar_existed = @new_cellar.exist?
 
-    if (@old_linked_keg = linked_old_linked_keg)
-      @old_linked_keg_record = old_linked_keg.linked_keg_record if old_linked_keg.linked?
-      @old_opt_record = old_linked_keg.opt_record if old_linked_keg.optlinked?
-      @new_linked_keg_record = HOMEBREW_CELLAR/"#{newname}/#{File.basename(old_linked_keg)}"
+    @old_linked_kegs = linked_old_linked_kegs
+    @old_full_linked_kegs = []
+    @old_opt_records = []
+    old_linked_kegs.each do |old_linked_keg|
+      @old_full_linked_kegs << old_linked_keg if old_linked_keg.linked?
+      @old_opt_records << old_linked_keg.opt_record if old_linked_keg.optlinked?
+    end
+    unless old_linked_kegs.empty?
+      @new_linked_keg_record = HOMEBREW_CELLAR/"#{newname}/#{File.basename(old_linked_kegs.first)}"
     end
 
     @old_pin_record = HOMEBREW_PINNED_KEGS/oldname
@@ -167,7 +171,7 @@ class Migrator
 
     new_tap = if old_tap
       old_tap_user, = old_tap.user
-      if (migrate_tap = old_tap.tap_migrations[formula.oldname])
+      if (migrate_tap = old_tap.tap_migrations[oldname])
         new_tap_user, new_tap_repo = migrate_tap.split("/")
         "#{new_tap_user}/#{new_tap_repo}"
       end
@@ -188,12 +192,12 @@ class Migrator
     end
   end
 
-  def linked_old_linked_keg
+  def linked_old_linked_kegs
     keg_dirs = []
     keg_dirs += new_cellar.subdirs if new_cellar.exist?
     keg_dirs += old_cellar.subdirs
     kegs = keg_dirs.map { |d| Keg.new(d) }
-    kegs.find(&:linked?) || kegs.find(&:optlinked?)
+    kegs.select { |keg| keg.linked? || keg.optlinked? }
   end
 
   def pinned?
@@ -209,7 +213,7 @@ class Migrator
     move_to_new_directory
     link_oldname_cellar
     link_oldname_opt
-    link_newname unless old_linked_keg.nil?
+    link_newname unless old_linked_kegs.empty?
     update_tabs
     return unless formula.outdated?
 
@@ -232,14 +236,14 @@ class Migrator
     unlock
   end
 
-  # Move everything from `Cellar/oldname` to `Cellar/newname`.
-  def move_to_new_directory
-    return unless old_cellar.exist?
+  def remove_conflicts(directory)
+    conflicted = T.let(false, T::Boolean)
 
-    if new_cellar.exist?
-      conflicted = T.let(false, T::Boolean)
-      old_cellar.each_child do |c|
-        next unless (new_cellar/c.basename).exist?
+    directory.each_child do |c|
+      if c.directory? && !c.symlink?
+        conflicted ||= remove_conflicts(c)
+      else
+        next unless (new_cellar/c.relative_path_from(old_cellar)).exist?
 
         begin
           FileUtils.rm_rf c
@@ -248,13 +252,36 @@ class Migrator
           onoe "#{new_cellar/c.basename} already exists."
         end
       end
+    end
 
-      odie "Remove #{new_cellar} manually and run `brew migrate #{oldname}`." if conflicted
+    conflicted
+  end
+
+  def merge_directory(directory)
+    directory.each_child do |c|
+      new_path = new_cellar/c.relative_path_from(old_cellar)
+
+      if c.directory? && !c.symlink? && new_path.exist?
+        merge_directory(c)
+        c.unlink
+      else
+        FileUtils.mv(c, new_path)
+      end
+    end
+  end
+
+  # Move everything from `Cellar/oldname` to `Cellar/newname`.
+  def move_to_new_directory
+    return unless old_cellar.exist?
+
+    if new_cellar.exist?
+      conflicted = remove_conflicts(old_cellar)
+      odie "Remove #{new_cellar} and #{old_cellar} manually and run `brew reinstall #{newname}`." if conflicted
     end
 
     oh1 "Moving #{Formatter.identifier(oldname)} versions to #{new_cellar}"
     if new_cellar.exist?
-      FileUtils.mv(old_cellar.children, new_cellar)
+      merge_directory(old_cellar)
     else
       FileUtils.mv(old_cellar, new_cellar)
     end
@@ -302,7 +329,7 @@ class Migrator
     # If old_keg wasn't linked then we just optlink a keg.
     # If old keg wasn't optlinked and linked, we don't call this method at all.
     # If formula is keg-only we also optlink it.
-    if formula.keg_only? || !old_linked_keg_record
+    if formula.keg_only? || old_full_linked_kegs.empty?
       begin
         new_keg.optlink(verbose: verbose?)
       rescue Keg::LinkError => e
@@ -340,10 +367,10 @@ class Migrator
 
   # Link keg to opt if it was linked before migrating.
   def link_oldname_opt
-    return unless old_opt_record
-
-    old_opt_record.delete if old_opt_record.symlink?
-    old_opt_record.make_relative_symlink(new_linked_keg_record)
+    old_opt_records.each do |old_opt_record|
+      old_opt_record.delete if old_opt_record.symlink?
+      old_opt_record.make_relative_symlink(new_linked_keg_record)
+    end
   end
 
   # After migration every `INSTALL_RECEIPT.json` has the wrong path to the formula
@@ -358,14 +385,16 @@ class Migrator
 
   # Remove `opt/oldname` link if it belongs to newname.
   def unlink_oldname_opt
-    return if old_opt_record.to_s.blank?
-    return unless old_opt_record.symlink?
-    return unless old_opt_record.exist?
     return unless new_linked_keg_record.exist?
-    return if new_linked_keg_record.realpath != old_opt_record.realpath
 
-    old_opt_record.unlink
-    old_opt_record.parent.rmdir_if_possible
+    old_opt_records.each do |old_opt_record|
+      next unless old_opt_record.symlink?
+      next unless old_opt_record.exist?
+      next if new_linked_keg_record.realpath != old_opt_record.realpath
+
+      old_opt_record.unlink
+      old_opt_record.parent.rmdir_if_possible
+    end
   end
 
   # Remove `Cellar/oldname` if it exists.
@@ -399,26 +428,25 @@ class Migrator
       new_cellar.subdirs.each do |d|
         newname_keg = Keg.new(d)
         newname_keg.unlink(verbose: verbose?)
-        newname_keg.uninstall if new_cellar_existed
+        newname_keg.uninstall unless new_cellar_existed
       end
     end
 
-    return if old_linked_keg.nil?
+    return if old_linked_kegs.empty?
 
     # The keg used to be linked and when we backup everything we restore
     # Cellar/oldname, the target also gets restored, so we are able to
     # create a keg using its old path
-    if old_linked_keg_record
-      begin
-        old_linked_keg.link(verbose: verbose?)
-      rescue Keg::LinkError
-        old_linked_keg.unlink(verbose: verbose?)
-        raise
-      rescue Keg::AlreadyLinkedError
-        old_linked_keg.unlink(verbose: verbose?)
-        retry
-      end
-    else
+    old_full_linked_kegs.each do |old_linked_keg|
+      old_linked_keg.link(verbose: verbose?)
+    rescue Keg::LinkError
+      old_linked_keg.unlink(verbose: verbose?)
+      raise
+    rescue Keg::AlreadyLinkedError
+      old_linked_keg.unlink(verbose: verbose?)
+      retry
+    end
+    (old_linked_kegs - old_full_linked_kegs).each do |old_linked_keg|
       old_linked_keg.optlink(verbose: verbose?)
     end
   end
